+++ CHANGELOG.md
@@ -8,8 +8,8 @@
   location
 - [#234](https://github.com/estruyf/vscode-demo-time/issues/234): Fix for negative values in the
   insert typing speed and improved the labels
-- [#235](https://github.com/estruyf/vscode-demo-time/issues/235): Added the `waitForNext` action to
-  pause the demo and wait until the `Demo Time: Continue Demo` command is triggered
+- [#235](https://github.com/estruyf/vscode-demo-time/issues/235): Added the `pause` action to pause
+  the demo and wait until the `Demo Time: Continue Demo` command is triggered
 
 ## [1.9.0] - 2025-08-25